+++ lamp/Genie/Genie/IO/BufferFile.hh
@@ -43,9 +43,9 @@ namespace Genie
 			
 			unsigned int SysPoll() const;
 			
-			int SysRead( char* data, std::size_t byteCount );
+			ssize_t SysRead( char* data, std::size_t byteCount );
 			
-			int SysWrite( const char* data, std::size_t byteCount );
+			ssize_t SysWrite( const char* data, std::size_t byteCount );
 			
 			off_t GetEOF() const;
